Refactor: Evaluation and plotting code moved to their own individual directories
